Move node copy functionality to Graph.Node
Not 100% sure if it's the correct thing to do, but at least
we keep the (de)serialization contained to one function/class.

The cypher map op seems to be ideal since we never have to specify
which values to copy, but the current solution is good enough imo.

We probably also have to copy the classification as far as the base
taxonomy is concerned.